+++ core/src/syscall/unix/usleep.rs
@@ -40,7 +40,7 @@ impl UsleepSyscall for NioUsleepSyscall {
             None => Duration::MAX,
         };
         if let Some(co) = crate::scheduler::SchedulableCoroutine::current() {
-            let syscall = crate::common::constants::Syscall::usleep;
+            let syscall = crate::common::constants::SyscallName::usleep;
             let new_state = crate::common::constants::SyscallState::Suspend(
                 crate::common::get_timeout_time(time),
             );